refactor(admin/semesters): SemesterRegisterDrawer sections 를 공용 Section 으로 교체
- 4개 section 파일이 shared/ui/Section 을 사용하도록 변경
- drawer 전용 SectionTitle 을 정의하던 shared.tsx 제거
